fifth attempt

+++ segregation_amplification/prior_versions/second_attempt/functions.py
@@ -34,7 +34,7 @@ def initialize_agents(spatial_shape, proportion_filled, type_bias, params):
     n_cells = spatial_shape[0] * spatial_shape[1]
 
     # Keep track of all the agents via IDs
-    agent_ids = np.arange(0, n_agents).astype(np.int)
+    agent_ids = np.arange(0, n_agents).astype(np.int64)
 
     # positional_indices records the 2-length r, c of each agent grid-cell based on their r, c.
     positional_indices = np.empty((spatial_shape[0], spatial_shape[1], 2), dtype=int)
@@ -60,11 +60,11 @@ def initialize_agents(spatial_shape, proportion_filled, type_bias, params):
     unoccupied_locations = shuffled_positional_indices[n_agents:]
 
     # Agent types are 1 = liberal, 2 = conservative. 0 is left blank to reflect absence of agents.
-    agent_types = np.random.randint(1, 3, size=n_agents).astype(np.int)
+    agent_types = np.random.randint(1, 3, size=n_agents).astype(np.int64)
 
     # agent ids and types are initialized as maps.
-    agent_ids_map = np.zeros((spatial_shape[0], spatial_shape[1]), dtype=np.int)
-    types_map = np.zeros((spatial_shape[0], spatial_shape[1]), dtype=np.int)
+    agent_ids_map = np.zeros((spatial_shape[0], spatial_shape[1]), dtype=np.int64)
+    types_map = np.zeros((spatial_shape[0], spatial_shape[1]), dtype=np.int64)
     for i in range(n_agents):
         agent_ids_map[shuffled_positional_indices[i, 0], shuffled_positional_indices[i, 1]] = i
         types_map[shuffled_positional_indices[i, 0], shuffled_positional_indices[i, 1]] = agent_types[i]
@@ -100,7 +100,7 @@ def get_masking_choice_full_resolve(spatial_shape, proportion_filled, game_type,
     average_reciprocal_response_from_masking = np.zeros(spatial_shape).astype(np.float32)
     agent_ids, agent_locations, unoccupied_locations, agent_types, agent_ids_map, types_map, d, s, r = \
         initialize_agents(spatial_shape, proportion_filled, type_bias, params)
-    neighborhood_radius = np.int(neighborhood_radius)
+    neighborhood_radius = np.int64(neighborhood_radius)
     n_iterations = 150
     for i in range(n_iterations):
